requirements.txt
tensorflow>=2.10.0
numpy
pillow
fastapi
uvicorn
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
